personal-ai-assistant/requirements.txt
# Production dependencies (pinned for reproducibility)
fastapi>=0.115.0,<1.0
uvicorn[standard]>=0.32.0,<1.0
pydantic>=2.0,<3
pydantic-settings>=2.0,<3
python-dotenv>=1.0,<2
python-multipart>=0.0.9

# LangChain / LangGraph (1.x stack; langchain-mcp-adapters 0.2+ needs langchain-core>=1)
langchain>=1.0,<2
langchain-openai>=1.0,<2
langchain-community>=0.4,<1
langchain-text-splitters>=1.0,<2
langchain-core>=1.0,<2
langgraph>=1.0,<2
langsmith>=0.2,<1

# Postgres: pgvector, checkpointer, store
psycopg[binary,pool]>=3.1,<4
psycopg2-binary>=2.9,<3
pgvector>=0.2,<1
langgraph-checkpoint-postgres>=3.0,<4

# Document loaders
pypdf>=5.0,<6
unstructured[md]>=0.16; sys_platform != "win32"

# MCP (0.2.x matches langchain-core 1.x)
langchain-mcp-adapters>=0.2,<0.3

# Optional: agent web search (DuckDuckGo, no API key)
duckduckgo-search>=6.0,<8
